refactor(useCouponSelected.ts): CouponList 컴포넌트에서 useCouponSelected 도메인 로직 분리
