+++ src/pages/checkout/components/CouponModal/CouponList.tsx
@@ -1,52 +1,20 @@
-import { useRecoilState } from 'recoil';
-import { activeCouponCodesState, couponSelectedState } from '@store/couponStore';
 import useCouponList from '@hooks/coupon/useCouponList';
-import { COUPON_POLICY } from '@constants/policy';
+import useCouponSelected from '@hooks/coupon/useCouponSelected';
 import CouponItem from './CouponItem';
 
 export default function CouponList() {
-  const [couponSelected, setCouponSelected] = useRecoilState(couponSelectedState);
-  const [activeCouponCodes, setActiveCouponCodes] = useRecoilState(activeCouponCodesState);
-
-  const isFulledActiveCoupons = activeCouponCodes.length === COUPON_POLICY.max_active_coupon_amount;
-
   const { availableCoupons, unAvailableCoupons } = useCouponList();
-
-  const handleToggleCouponCheckbox = (toggledCouponCode: string) => {
-    const newCheckedState = !couponSelected[toggledCouponCode];
-
-    if (
-      newCheckedState === true &&
-      activeCouponCodes.length >= COUPON_POLICY.max_active_coupon_amount
-    )
-      return;
-
-    const newCouponSelected = {
-      ...couponSelected,
-      [toggledCouponCode]: newCheckedState,
-    };
-    setCouponSelected(newCouponSelected);
-
-    if (newCheckedState) {
-      const newActiveCoupons = [...activeCouponCodes, toggledCouponCode];
-      setActiveCouponCodes(newActiveCoupons);
-    } else {
-      const newActiveCoupons = activeCouponCodes.filter((code) => code !== toggledCouponCode);
-      setActiveCouponCodes(newActiveCoupons);
-    }
-  };
+  const { couponSelected, isDisableCoupon, handleToggleCouponCheckbox } = useCouponSelected();
 
   return (
     <>
       {availableCoupons.map((coupon) => {
-        const includeActiveCoupon = activeCouponCodes.includes(coupon.code);
-        const isDisableCoupon = isFulledActiveCoupons && !includeActiveCoupon;
         return (
           <CouponItem
             key={coupon.code}
             coupon={coupon}
             isChecked={couponSelected[coupon.code]}
-            isDisableCoupon={isDisableCoupon}
+            isDisableCoupon={isDisableCoupon(coupon.code)}
             onChange={() => {
               handleToggleCouponCheckbox(coupon.code);
             }}
